Moved files around and updated cmake
