+++ src/engine.cpp
@@ -1,5 +1,6 @@
-#include "engine.hpp"
+#include "engine.h"
 
+// Engine Lifecycle:
 void Engine::Engine::on_create() {
     initscr();
     cbreak();
@@ -21,6 +22,11 @@ void Engine::Engine::on_destroy() {
 }
 
 Engine::Engine::Engine() {}
+
+// Engine Utilities:
+void Engine::Engine::draw_to_window(int column, int row, char character) {
+    mvwaddch(p_window, column, row, character);
+}
 
 void Engine::print_2D_vector(std::vector<std::vector<char>> p_matrix) {
     for (std::size_t column = 0; column < p_matrix.size(); ++column) {
