+++ src/strings.h
